Move createCommitmentMessageRelay to communication module
