+++ packages/wallet/src/redux/protocols/direct-funding/reducer.ts
@@ -14,9 +14,9 @@ import {
 } from '../transaction-submission/reducer';
 import { isTerminal, isSuccess } from '../transaction-submission/states';
 import * as states from './state';
-import { createCommitmentMessageRelay } from '../reducer-helpers';
 import { theirAddress } from '../../channel-store';
 import * as channelStoreReducer from '../../channel-store/reducer';
+import { createCommitmentMessageRelay } from '../../../communication';
 
 type DFReducer = ProtocolReducer<states.DirectFundingState>;
