+++ packages/wallet/src/redux/protocols/indirect-funding/player-a/reducer.ts
@@ -17,7 +17,6 @@ import { CommitmentType, Commitment, getChannelId } from '../../../../domain';
 import { Channel } from 'fmg-core/lib/channel';
 import { CONSENSUS_LIBRARY_ADDRESS } from '../../../../constants';
 import { getChannel, theirAddress } from '../../../channel-store';
-import { createCommitmentMessageRelay } from '../../reducer-helpers';
 import { DirectFundingAction } from '../../direct-funding';
 import { directFundingRequested } from '../../direct-funding/actions';
 import {
@@ -28,6 +27,7 @@ import {
 } from '../../direct-funding/state';
 import { directFundingStateReducer } from '../../direct-funding/reducer';
 import { addHex } from '../../../../utils/hex-utils';
+import { createCommitmentMessageRelay } from '../../../../communication';
 
 type ReturnVal = ProtocolStateWithSharedData<IndirectFundingState>;
 type IDFAction = actions.indirectFunding.Action;
